Move protocol specific neighbor handling to the protocol specific plugins
